Struct tough::editor::signed::SignedDelegatedTargets

pub struct SignedDelegatedTargets { /* fields omitted */ }

A set of signed targets role metadata.

Implementations

impl SignedDelegatedTargets

pub fn write<P>(&self, outdir: P, consistent_snapshot: bool) -> Result<()> where
P: AsRef<Path>,

Writes the metadata to the given directory. If consistent snapshots are used, the appropriate files are prefixed with their version.

pub fn roles(self) -> Vec<SignedRole<DelegatedTargets>>

Returns all SignedRole<DelegatedTargets>> contained by this SignedDelegatedTargets

pub fn link_targets<P1, P2>(
&self,
indir: P1,
outdir: P2,
replace_behavior: PathExists
) -> Result<()> where
P1: AsRef<Path>,
P2: AsRef<Path>,

Crawls a given directory and symlinks any targets found to the given “out” directory. If consistent snapshots are used, the target files are prefixed with their sha256.

For each file found in the indir, the method gets the filename and if the filename exists in Targets, the file’s sha256 is compared against the data in Targets. If this data does not match, the method will fail.

pub fn copy_targets<P1, P2>(
&self,
indir: P1,
outdir: P2,
replace_behavior: PathExists
) -> Result<()> where
P1: AsRef<Path>,
P2: AsRef<Path>,

Crawls a given directory and copies any targets found to the given “out” directory. If consistent snapshots are used, the target files are prefixed with their sha256.

For each file found in the indir, the method gets the filename and if the filename exists in Targets, the file’s sha256 is compared against the data in Targets. If this data does not match, the method will fail.

pub fn link_target(
&self,
input_path: &Path,
outdir: &Path,
replace_behavior: PathExists,
target_filename: Option<&TargetName>
) -> Result<()>

Symlinks a single target to the desired directory. If target_filename is given, it becomes the filename suffix, otherwise the original filename is used. (A unique filename prefix is used if consistent snapshots are enabled.) Fails if the target already exists in the repo with a different hash, or if it has the same hash but is not a symlink. Using the replace_behavior parameter, you can decide what happens if it exists with the same hash and file type - skip, fail, or replace.

pub fn copy_target(
&self,
input_path: &Path,
outdir: &Path,
replace_behavior: PathExists,
target_filename: Option<&TargetName>
) -> Result<()>

Copies a single target to the desired directory. If target_filename is given, it becomes the filename suffix, otherwise the original filename is used. (A unique filename prefix is used if consistent hashing is enabled.) Fails if the target already exists in the repo with a different hash, or if it has the same hash but is not a regular file. Using the replace_behavior parameter, you can decide what happens if it exists with the same hash and file type - skip, fail, or replace.