Struct tough::editor::targets::TargetsEditor

pub struct TargetsEditor { /* fields omitted */ }

If you are not working with a repository that utilizes delegated targets, use the RepositoryEditor.

TargetsEditor contains the various bits of data needed to construct or edit a Targets role.

TargetsEditor should only be used with repositories that utilize delegated targets. TargetsEditor cannot create a SignedRepository. Whenever a user has access to snapshot.json and timestamp.json keys, RepositoryEditor should be used so that a SignedRepository can be created.

A new Targets may be created using the new() method.

An existing Targets may be loaded and edited using the from_targets() method. When a targets is loaded in this way, versions and expirations are discarded. It is good practice to update these whenever a repo is changed.

A Targets from an existing repository can be loaded using the from_repo() method. Targets loaded this way will have the versions and expirations removed, but the proper keyholder to sign the targets and the Transport used to load the repo will be saved.

Targets, versions, and expirations may be added to their respective roles via the provided “setter” methods. The final step in the process is the sign() method, which takes a given set of signing keys, builds each of the roles using the data provided, and signs the roles. This results in a SignedDelegatedTargets which can be used to write the updated metadata to disk.

Implementations

impl TargetsEditor

pub fn new(name: &str) -> Self

Creates a TargetsEditor for a newly created role

pub fn from_targets(name: &str, targets: Targets, key_holder: KeyHolder) -> Self

Creates a TargetsEditor with the provided targets and keyholder version and expires are thrown out to encourage updating the version and expiration

pub fn from_repo(repo: Repository, name: &str) -> Result<Self>

Creates a TargetsEditor with the provided targets from an already loaded repo version and expires are thrown out to encourage updating the version and expiration If a Repository has been loaded, use from_repo() to preserve the Transport and Limits.

pub fn limits(&mut self, limits: Limits)

Adds limits to the TargetsEditor, only necessary if loading a role

pub fn transport(&mut self, transport: Box<dyn Transport>)

Add a transport to the TargetsEditor, only necessary if loading a role

pub fn add_target<T, E>(&mut self, name: T, target: Target) -> Result<&mut Self> where
T: TryInto<TargetName, Error = E>,
E: Display,

Add a Target to the Targets role

pub fn add_target_path<P>(&mut self, target_path: P) -> Result<&mut Self> where
P: AsRef<Path>,

Add a target to the repository using its path

Note: This function builds a Target synchronously; no multithreading or parallelism is used. If you have a large number of targets to add, and require advanced performance, you may want to construct Targets directly in parallel and use add_target().

pub fn add_target_paths<P>(&mut self, targets: Vec<P>) -> Result<&mut Self> where
P: AsRef<Path>,

Add a list of target paths to the targets

See the note on add_target_path() regarding performance.

pub fn remove_target(&mut self, name: &TargetName) -> &mut Self

Remove a Target from the targets if it exists

pub fn clear_targets(&mut self) -> &mut Self

Remove all targets from this role

pub fn version(&mut self, version: NonZeroU64) -> &mut Self

Set the version

pub fn expires(&mut self, expires: DateTime<Utc>) -> &mut Self

Set the expiration

pub fn add_key(
&mut self,
keys: HashMap<Decoded<Hex>, Key>,
role: Option<&str>
) -> Result<&mut Self>

Adds a key to delegations keyids, adds the key to role if it is provided

pub fn remove_key(
&mut self,
keyid: &Decoded<Hex>,
role: Option<&str>
) -> Result<&mut Self>

Removes a key from delegations keyids, if a role is specified the key is only removed from the role

pub fn delegate_role(
&mut self,
targets: Signed<DelegatedTargets>,
paths: PathSet,
key_pairs: HashMap<Decoded<Hex>, Key>,
keyids: Vec<Decoded<Hex>>,
threshold: NonZeroU64
) -> Result<&mut Self>

Adds a DelegatedRole to new_roles To use delegate_role() a new Targets should be created using TargetsEditor::new() followed by create_signed() to provide a Signed<DelegatedTargets> for the new role.

pub fn remove_role(&mut self, role: &str, recursive: bool) -> Result<&mut Self>

Removes a role from delegations If recursive is false, role is only removed if it is directly delegated by this role If true removes whichever role eventually delegates ‘role’

pub fn add_role(
&mut self,
name: &str,
metadata_url: &str,
paths: PathSet,
threshold: NonZeroU64,
keys: Option<HashMap<Decoded<Hex>, Key>>
) -> Result<&mut Self>

Adds a role to new_roles using a metadata file located at metadata_url/name.json add_role() uses delegate_role() to add a role from an existing metadata file.

pub fn build_targets(&self) -> Result<DelegatedTargets>

Build the Targets struct Adds in the new roles and new targets

pub fn create_signed(
&self,
keys: &[Box<dyn KeySource>]
) -> Result<Signed<DelegatedTargets>>

Creates a Signed<DelegatedTargets> for only this role using the provided keys This is used to create a Signed<DelegatedTargets> for the role instead of a SignedDelegatedTargets like sign() creates. SignedDelegatedTargets can contain more than 1 Signed<DelegatedTargets> create_signed() guarantees that only 1 Signed<DelegatedTargets> is created and that it is the one representing the current targets. create_signed() should be used whenever the result of TargetsEditor is not being written.

pub fn sign(
&self,
keys: &[Box<dyn KeySource>]
) -> Result<SignedDelegatedTargets>

Creates a SignedDelegatedTargets for the Targets role being edited and all added roles If key_holder was not assigned then this is a newly created role and needs to be signed with a custom delegations as its key_holder