Enum tough::error::Error

#[non_exhaustive]
pub enum Error {

    AbsolutePath {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DatastoreInit {
        source: Error,
        backtrace: Backtrace,
    },
    DatastoreCreate {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DatastoreOpen {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DatastoreRemove {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DatastoreSerialize {
        what: String,
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DirCreate {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    ExpiredMetadata {
        role: RoleType,
        backtrace: Backtrace,
    },
    FileMetadata {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    FileOpen {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    FileRead {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    FileParseJson {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    FileUrl {
        path: PathBuf,
        backtrace: Backtrace,
    },
    FileWrite {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    HashMismatch {
        context: String,
        calculated: String,
        expected: String,
        backtrace: Backtrace,
    },
    InvalidFileType {
        path: PathBuf,
        backtrace: Backtrace,
    },
    InvalidTargetName {
        inner: String,
        backtrace: Backtrace,
    },
    JoinUrl {
        path: String,
        url: Url,
        source: ParseError,
        backtrace: Backtrace,
    },
    JoinUrlEncoded {
        original: String,
        encoded: String,
        filename: String,
        url: Url,
        source: ParseError,
        backtrace: Backtrace,
    },
    KeyPairFromKeySource {
        source: Box<dyn Error + Send + Sync + 'static>,
        backtrace: Backtrace,
    },
    KeyRejected {
        source: KeyRejected,
        backtrace: Backtrace,
    },
    KeysNotFoundInRoot {
        backtrace: Backtrace,
    },
    KeyUnrecognized {
        backtrace: Backtrace,
    },
    LinkCreate {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    MaxSizeExceeded {
        max_size: u64,
        specifier: &'static str,
        backtrace: Backtrace,
    },
    MaxUpdatesExceeded {
        max_root_updates: u64,
        backtrace: Backtrace,
    },
    MetaMissing {
        file: &'static str,
        role: RoleType,
        backtrace: Backtrace,
    },
    Missing {
        field: String,
        backtrace: Backtrace,
    },
    NamedTempFileCreate {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    NamedTempFilePersist {
        path: PathBuf,
        source: PersistError,
        backtrace: Backtrace,
    },
    NoFileName {
        path: PathBuf,
        backtrace: Backtrace,
    },
    NoRoleKeysinRoot {
        role: String,
    },
    OlderMetadata {
        role: RoleType,
        current_version: u64,
        new_version: u64,
        backtrace: Backtrace,
    },
    ParseMetadata {
        role: RoleType,
        source: Error,
        backtrace: Backtrace,
    },
    ParseTrustedMetadata {
        source: Error,
        backtrace: Backtrace,
    },
    ParseUrl {
        url: String,
        source: ParseError,
        backtrace: Backtrace,
    },
    PathExistsFail {
        path: PathBuf,
        backtrace: Backtrace,
    },
    PathIsNotFile {
        path: PathBuf,
        backtrace: Backtrace,
    },
    PathIsNotTarget {
        path: PathBuf,
        backtrace: Backtrace,
    },
    PathUtf8 {
        path: PathBuf,
        backtrace: Backtrace,
    },
    RemoveTarget {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    SaveTargetDirInfo {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    SaveTargetOutdir {
        path: PathBuf,
        backtrace: Backtrace,
    },
    SaveTargetOutdirCanonicalize {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    SaveTargetNoParent {
        path: PathBuf,
        name: TargetName,
        backtrace: Backtrace,
    },
    SaveTargetNotFound {
        name: TargetName,
        backtrace: Backtrace,
    },
    SaveTargetUnsafePath {
        name: TargetName,
        outdir: PathBuf,
        filepath: PathBuf,
    },
    SerializeRole {
        role: String,
        source: Error,
        backtrace: Backtrace,
    },
    SerializeSignedRole {
        role: String,
        source: Error,
        backtrace: Backtrace,
    },
    Sign {
        source: Unspecified,
        backtrace: Backtrace,
    },
    SignMessage {
        source: Box<dyn Error + Send + Sync + 'static>,
        backtrace: Backtrace,
    },
    SigningKeysNotFound {
        role: String,
    },
    SpecVersion {
        given: String,
        supported: String,
        backtrace: Backtrace,
    },
    SystemTimeSteppedBackward {
        sys_time: DateTime<Utc>,
        latest_known_time: DateTime<Utc>,
    },
    TargetFileTypeMismatch {
        expected: String,
        found: String,
        path: PathBuf,
        backtrace: Backtrace,
    },
    TargetFromPath {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    TargetNameResolve {
        name: String,
        source: Error,
    },
    TargetNameComponentsEmpty {
        name: String,
    },
    TargetNameRootMissing {
        name: String,
    },
    Transport {
        url: Url,
        source: TransportError,
        backtrace: Backtrace,
    },
    UnsafeTargetNameDotDot {},
    UnsafeTargetNameEmpty {
        name: String,
    },
    UnsafeTargetNameSlash {
        name: String,
    },
    VerifyMetadata {
        role: RoleType,
        source: Error,
        backtrace: Backtrace,
    },
    VerifyRoleMetadata {
        role: String,
        source: Error,
        backtrace: Backtrace,
    },
    VerifyTrustedMetadata {
        source: Error,
        backtrace: Backtrace,
    },
    VersionMismatch {
        role: RoleType,
        fetched: u64,
        expected: u64,
        backtrace: Backtrace,
    },
    CacheFileRead {
        url: Url,
        source: Error,
        backtrace: Backtrace,
    },
    CacheFileWrite {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    CacheDirectoryCreate {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    CacheTargetWrite {
        path: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    CacheTargetMissing {
        target_name: TargetName,
        source: Error,
        backtrace: Backtrace,
    },
    WalkDir {
        directory: PathBuf,
        source: Error,
        backtrace: Backtrace,
    },
    DelegateNotFound {
        name: String,
    },
    TargetsNotFound {
        name: String,
        source: Error,
    },
    DelegateMissing {
        name: String,
        source: Error,
    },
    NoTargets,
    NoDelegations,
    DelegatedRolesNotConsistent {
        name: String,
    },
    InvalidPath {
        source: Error,
    },
    RoleNotInMeta {
        name: String,
    },
    KeyNotFound {
        role: String,
        source: Error,
        backtrace: Backtrace,
    },
    NoKeys {
        role: String,
    },
    InvalidInto {
        source: TryFromIntError,
        backtrace: Backtrace,
    },
    InvalidThreshold {
        backtrace: Backtrace,
    },
    JsonSerialization {
        source: Error,
        backtrace: Backtrace,
    },
    InvalidPathPermission {
        name: String,
        paths: Vec<String>,
        source: Error,
    },
    InvalidRoleCount {
        count: usize,
    },
    TargetsMap {
        source: Error,
    },
    NoKeyHolder,
    MissingLimits,
    MissingTransport,
    UnstableRoot {
        role: RoleType,
        actual: usize,
        threshold: u64,
    },
    TargetsEditorSome,

}

The error type for this library.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AbsolutePath

Fields

path: PathBuf

source: Error

backtrace: Backtrace

DatastoreInit

Fields

source: Error

backtrace: Backtrace

DatastoreCreate

Fields

path: PathBuf

source: Error

backtrace: Backtrace

The library failed to create a file in the datastore.

DatastoreOpen

Fields

path: PathBuf

source: Error

backtrace: Backtrace

The library failed to open a file in the datastore.

DatastoreRemove

Fields

path: PathBuf

source: Error

backtrace: Backtrace

The library failed to remove a file in the datastore.

DatastoreSerialize

Fields

what: String

path: PathBuf

source: Error

backtrace: Backtrace

The library failed to serialize an object to JSON to the datastore.

DirCreate

Fields

path: PathBuf

source: Error

backtrace: Backtrace

ExpiredMetadata

Fields

role: RoleType

backtrace: Backtrace

A metadata file has expired.

FileMetadata

Fields

path: PathBuf

source: Error

backtrace: Backtrace

FileOpen

Fields

path: PathBuf

source: Error

backtrace: Backtrace

FileRead

Fields

path: PathBuf

source: Error

backtrace: Backtrace

FileParseJson

Fields

path: PathBuf

source: Error

backtrace: Backtrace

FileUrl

Fields

path: PathBuf

backtrace: Backtrace

FileWrite

Fields

path: PathBuf

source: Error

backtrace: Backtrace

HashMismatch

Fields

context: String

calculated: String

expected: String

backtrace: Backtrace

A downloaded target’s checksum does not match the checksum listed in the repository metadata.

InvalidFileType

Fields

path: PathBuf

backtrace: Backtrace

InvalidTargetName

Fields

inner: String

backtrace: Backtrace

JoinUrl

Fields

path: String

url: Url

source: ParseError

backtrace: Backtrace

The library failed to create a URL from a base URL and a path.

JoinUrlEncoded

Fields

original: String

encoded: String

filename: String

url: Url

source: ParseError

backtrace: Backtrace

KeyPairFromKeySource

Fields

source: Box<dyn Error + Send + Sync + 'static>

backtrace: Backtrace

KeyRejected

Fields

source: KeyRejected

backtrace: Backtrace

KeysNotFoundInRoot

Fields

backtrace: Backtrace

KeyUnrecognized

Fields

backtrace: Backtrace

LinkCreate

Fields

path: PathBuf

source: Error

backtrace: Backtrace

MaxSizeExceeded

Fields

max_size: u64

specifier: &'static str

backtrace: Backtrace

A file’s maximum size exceeded a limit set by the consumer of this library or the metadata.

MaxUpdatesExceeded

Fields

max_root_updates: u64

backtrace: Backtrace

The maximum root updates setting was exceeded.

MetaMissing

Fields

file: &'static str

role: RoleType

backtrace: Backtrace

A required reference to a metadata file is missing from a metadata file.

Missing

Fields

field: String

backtrace: Backtrace

NamedTempFileCreate

Fields

path: PathBuf

source: Error

backtrace: Backtrace

NamedTempFilePersist

Fields

path: PathBuf

source: PersistError

backtrace: Backtrace

NoFileName

Fields

path: PathBuf

backtrace: Backtrace

Unable to determine file name (path ends in ‘..’ or is ‘/’)

NoRoleKeysinRoot

Fields

role: String

OlderMetadata

Fields

role: RoleType

current_version: u64

new_version: u64

backtrace: Backtrace

A downloaded metadata file has an older version than a previously downloaded metadata file.

ParseMetadata

Fields

role: RoleType

source: Error

backtrace: Backtrace

The library failed to parse a metadata file, either because it was not valid JSON or it did not conform to the expected schema.

Invalid JSON errors read like:

EOF while parsing a string at line 1 column 14

Schema non-conformance errors read like:

invalid type: integer 2, expected a string at line 1 column 11

missing field sig at line 1 column 16

ParseTrustedMetadata

Fields

source: Error

backtrace: Backtrace

The library failed to parse the trusted root metadata file, either because it was not valid JSON or it did not conform to the expected schema. The trusted root metadata file is the file is either the root argument passed to Repository::load, or the most recently cached and validated root metadata file.

ParseUrl

Fields

url: String

source: ParseError

backtrace: Backtrace

Failed to parse a URL provided to Repository::load.

PathExistsFail

Fields

path: PathBuf

backtrace: Backtrace

PathIsNotFile

Fields

path: PathBuf

backtrace: Backtrace

PathIsNotTarget

Fields

path: PathBuf

backtrace: Backtrace

PathUtf8

Fields

path: PathBuf

backtrace: Backtrace

Path isn’t a valid UTF8 string

RemoveTarget

Fields

path: PathBuf

source: Error

backtrace: Backtrace

SaveTargetDirInfo

Fields

path: PathBuf

source: Error

backtrace: Backtrace

SaveTargetOutdir

Fields

path: PathBuf

backtrace: Backtrace

SaveTargetOutdirCanonicalize

Fields

path: PathBuf

source: Error

backtrace: Backtrace

SaveTargetNoParent

Fields

path: PathBuf

name: TargetName

backtrace: Backtrace

SaveTargetNotFound

Fields

name: TargetName

backtrace: Backtrace

SaveTargetUnsafePath

Fields

name: TargetName

outdir: PathBuf

filepath: PathBuf

SerializeRole

Fields

role: String

source: Error

backtrace: Backtrace

SerializeSignedRole

Fields

role: String

source: Error

backtrace: Backtrace

Sign

Fields

source: Unspecified

backtrace: Backtrace

SignMessage

Fields

source: Box<dyn Error + Send + Sync + 'static>

backtrace: Backtrace

SigningKeysNotFound

Fields

role: String

SpecVersion

Fields

given: String

supported: String

backtrace: Backtrace

SystemTimeSteppedBackward

Fields

sys_time: DateTime<Utc>

latest_known_time: DateTime<Utc>

System time is behaving irrationally, went back in time

TargetFileTypeMismatch

Fields

expected: String

found: String

path: PathBuf

backtrace: Backtrace

TargetFromPath

Fields

path: PathBuf

source: Error

backtrace: Backtrace

TargetNameResolve

Fields

name: String

source: Error

TargetNameComponentsEmpty

Fields

name: String

TargetNameRootMissing

Fields

name: String

Transport

Fields

url: Url

source: TransportError

backtrace: Backtrace

A transport error occurred while fetching a URL.

UnsafeTargetNameDotDot

Fields

UnsafeTargetNameEmpty

Fields

name: String

UnsafeTargetNameSlash

Fields

name: String

VerifyMetadata

Fields

role: RoleType

source: Error

backtrace: Backtrace

A metadata file could not be verified.

VerifyRoleMetadata

Fields

role: String

source: Error

backtrace: Backtrace

VerifyTrustedMetadata

Fields

source: Error

backtrace: Backtrace

The trusted root metadata file could not be verified.

VersionMismatch

Fields

role: RoleType

fetched: u64

expected: u64

backtrace: Backtrace

A fetched metadata file did not have the version we expected it to have.

CacheFileRead

Fields

url: Url

source: Error

backtrace: Backtrace

CacheFileWrite

Fields

path: PathBuf

source: Error

backtrace: Backtrace

CacheDirectoryCreate

Fields

path: PathBuf

source: Error

backtrace: Backtrace

CacheTargetWrite

Fields

path: PathBuf

source: Error

backtrace: Backtrace

CacheTargetMissing

Fields

target_name: TargetName

source: Error

backtrace: Backtrace

WalkDir

Fields

directory: PathBuf

source: Error

backtrace: Backtrace

DelegateNotFound

Fields

name: String

TargetsNotFound

Fields

name: String

source: Error

DelegateMissing

Fields

name: String

source: Error

NoTargets

NoDelegations

DelegatedRolesNotConsistent

Fields

name: String

InvalidPath

Fields

source: Error

Target doesn’t have proper permissions from parent delegations

RoleNotInMeta

Fields

name: String

KeyNotFound

Fields

role: String

source: Error

backtrace: Backtrace

NoKeys

Fields

role: String

InvalidInto

Fields

source: TryFromIntError

backtrace: Backtrace

InvalidThreshold

Fields

backtrace: Backtrace

JsonSerialization

Fields

source: Error

backtrace: Backtrace

The library failed to serialize an object to JSON.

InvalidPathPermission

Fields

name: String

paths: Vec<String>

source: Error

Invalid path permissions

InvalidRoleCount

Fields

count: usize

SignedDelegatedTargets has more than 1 signed targets

TargetsMap

Fields

source: Error

Could not create a targets map

NoKeyHolder

A key_holder wasn’t set

MissingLimits

MissingTransport

UnstableRoot

Fields

role: RoleType

actual: usize

threshold: u64

Root creates an unloadable repo

TargetsEditorSome