Trait tough::schema::Role

pub trait Role: Serialize {
    const TYPE: RoleType;

    fn expires(&self) -> DateTime<Utc>;

    fn version(&self) -> NonZeroU64;

    fn filename(&self, consistent_snapshot: bool) -> String;

    fn role_id(&self) -> RoleId { ... }

    fn canonical_form(&self) -> Result<Vec<u8>> { ... }
}

Common trait implemented by all roles.

Associated Constants

const TYPE: RoleType

The type of role this object represents.

Required methods

fn expires(&self) -> DateTime<Utc>

Determines when metadata should be considered expired and no longer trusted by clients.

fn version(&self) -> NonZeroU64

An integer that is greater than 0. Clients MUST NOT replace a metadata file with a version number less than the one currently trusted.

fn filename(&self, consistent_snapshot: bool) -> String

The filename that the role metadata should be written to

Provided methods

fn role_id(&self) -> RoleId

The RoleId corresponding to the role

fn canonical_form(&self) -> Result<Vec<u8>>

A deterministic JSON serialization used when calculating the digest of a metadata object. More info on canonical JSON