Trait tough::sign::Sign

pub trait Sign: Sync + Send {
    fn tuf_key(&self) -> Key;

    fn sign(
        &self, 
        msg: &[u8], 
        rng: &dyn SecureRandom
    ) -> Result<Vec<u8>, Box<dyn Error + Send + Sync + 'static>>;
}

This trait must be implemented for each type of key with which you will sign things.

Required methods

fn tuf_key(&self) -> Key

Returns the decoded key along with its scheme and other metadata

fn sign(
&self,
msg: &[u8],
rng: &dyn SecureRandom
) -> Result<Vec<u8>, Box<dyn Error + Send + Sync + 'static>>

Signs the supplied message